remove mongodb & plant api to router
