+++ package.json
@@ -49,6 +49,7 @@
     "eslint-plugin-jsx-a11y": "^1.2.0",
     "eslint-plugin-react": "^5.1.1",
     "extract-text-webpack-plugin": "^1.0.1",
+    "http-proxy-middleware": "^0.16.0",
     "proxy-middleware": "^0.15.0",
     "socket.io": "^1.4.6",
     "style-loader": "^0.13.1",
